User limit stream builder

+++ lora_monitor/lib/domain/user_limit.dart
@@ -1,12 +1,12 @@
 class UserLimit {
-  int max;
-  int min;
+  double max;
+  double min;
   String measure;
 
   UserLimit(this.max, this.min, this.measure);
 
   factory UserLimit.fromJson(Map<dynamic, dynamic> json) => UserLimit(
-      json['max'] as int, json['min'] as int, json['measure'] as String);
+      json['max'] as double, json['min'] as double, json['measure'] as String);
 
   Map<dynamic, dynamic> toJson() => {
         'max': max,
